backend/requirements.txt
# Django and related packages
django
djangorestframework
dj-rest-auth[with_social]
django-allauth
django-cors-headers

# Environment variables
python-dotenv

# Image support
Pillow

# Encryption
cryptography
